rename a module
